+++ src/utils/types.ts
@@ -1,4 +1,0 @@
-export interface Preferences {
-  auth: string;
-  baseUrl?: string;
-}