+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # GitHub Changelog
 
+## [Repo actions] - 2023-05-18
+
+- add: Copy `Projj Add Command` to Clipboard
+
 ## [Initial Version] - 2023-05-17